Try to deal with parallel deletion of files on slaves

+++ src/slave/src/org/drftpd/slave/Slave.java
@@ -487,13 +487,19 @@ public class Slave {
 					if (dir.getPath().length() <= root.getPath().length()) {
 						break;
 					}
-
-					java.io.File tmpFile = dir.getParentFile();
-
-					if (!dir.delete()) {
-						throw new PermissionDeniedException("delete failed on " + path);
+					
+					if(!dir.exists()) {
+						logger.info("Parent dir was empty, but doesn't exist anymore, that is fine " + dir.getPath());
+						break;
 					}
+					
+					java.io.File tmpFile = dir.getParentFile();
 					logger.info("Dir empty, rmdir: " + dir.getPath());
+
+					
+					if (!dir.delete()) {
+						throw new PermissionDeniedException("delete of parent dir failed on " + path);
+					}
 
 					if (tmpFile == null) {
 						break;
